+++ client/src/components/screens/ProfilePage/LogOutButton.tsx
@@ -1,8 +1,8 @@
 import { useRouter } from "next/router"
 import React from "react"
-import { useAppDispatch } from "../hooks/useAppDispatch"
-import $api from "../http"
-import { clearUser, setIsAuth, setUser } from "../redux/reducers/userSlice"
+import { useAppDispatch } from "hooks/useAppDispatch"
+import $api from "http"
+import { clearUser, setIsAuth, setUser } from "redux/reducers/userSlice"
 
 
 const LogOutButton: React.FC = () => {
